Poison stacks, ranged attack tweak
Poison effect now deals damage relative to number of stacks
Ranged attack only takes status effect from enemies if it has none

+++ ProjectGreedFallenKingdom/Assets/Scripts/Entities/Enemy/EnemyHealth.cs
@@ -57,7 +57,7 @@ public class EnemyHealth : MonoBehaviour
     }
 
     //======================================================================
-    public void UpdateCurrentHealth(int amount = 0)
+    public void UpdateCurrentHealth(float amount = 0)
     {
         if (amount != 0) {
                 currentHealth += amount;
